Stop juggling OS X Emacs in .bashrc
Now doing it just with files in ~/bin.  Was moved to do this by
org-mode which tries to set EMACS in the environment and then compile
with it, rude awakening when OS X 10.6 Emacs 22 used to byte compile
files that I then try to load into Emacs 23.  ~/bin scripts seem less
invasive.  I may check them in eventually, but they might be
arch-specific.