Dispatch media keys to VLC from Hammerspoon
VLC wants to create an event tap to listen for media keys, which means
granting accessibility, and giving VLC that permission seems risky to
me.  Hammerspoon already has accessibility (and it's not playing
random videos I download), so I just added the ability to tell VLC
what to do in my Hammerspoon config.