Replace fasd with zoxide
Proximate motivation for this was noticing that commands such as
`blahblah /some/dir/with/lots/of/files/*` took a *very* long time to
run.  Tracking this down was kind of hard, but I eventually realized
it was fasd parsing every command line argument, like twice, in a
combination of ugly POSIX shell and Awk.  Worse yet, I think it was
doing it for features I don't use or necessarily need.  As I started
reviewing all the alternatives (see comment in `.bashrc` for my top
contenders), I noticed several of them were doing a lot of work in
shell and Awk.  zoxide is basically all Rust, so hopefully I won't
have speed problems in the future, and I probably gained some useful
functionality, like the `zi` alias.