Add macOS Dark Mode theme changing via Hammerspoon
This lets me alter my Emacs, iTerm, and Textual themes to match light
vs. dark mode in macOS.  This is probably not going to work right out
of the box on a new system, such as because of the lack of a Python
runtime installed in iTerm2, or because I haven't set my Textual
preferences to use this "Auto" theme correctly; I need to update my
Ansible playbooks.