Enable alternate screen in tmux
I'm sure I turned this off because it pisses me off and fucks up
scroll back 99% of the time when it's used.  However, now I want to
use yazi, a cool terminal file manager, but it really needs alternate
screen, otherwise your screen looks like 1,000% ass when it exits.
It's been fine for me for weeks on my work machine.